% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geom_dot_gradient.R
\name{ggplot_add.geom_dot_gradient}
\alias{ggplot_add.geom_dot_gradient}
\title{Add dot summaries with alpha gradients below violin plots}
\usage{
\method{ggplot_add}{geom_dot_gradient}(object, plot, object_name)
}
\arguments{
\item{object}{An object created by \code{\link[=geom_dot_gradient]{geom_dot_gradient()}}.}

\item{plot}{A ggplot object or a patchwork object.}

\item{object_name}{Not used.}
}
\value{
A ggplot object or a patchwork object with summary dots.
}
\description{
S3 method for adding a \code{"geom_dot_gradient"} object to a ggplot.
}
